Agregue el metodos de inseertar productos a medias

+++ modelo/Mproductos.cs
@@ -4,7 +4,7 @@
     {
         public long? id { get; set; }
         public string? descripcion { get; set; }
-        public decimal? precio { get; set;}
+        public long? precio { get; set;}
 
     }
 }
